requirements.txt
chromadb
sentence-transformers
fastapi
uvicorn
python-dotenv
tree_sitter
tree_sitter_cpp
llama-cpp-python
